requirements.txt
absl-py==1.4.0
aiosignal==1.3.1
astunparse==1.6.3
attrs==23.1.0
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.2.1
contourpy==1.0.7
cycler==0.11.0
decorator==5.1.1
distlib==0.3.6
dm-tree==0.1.8
Farama-Notifications==0.0.4
filelock==3.12.0
flatbuffers==1.12
fonttools==4.39.4
frozenlist==1.3.3
fsspec==2023.6.0
gast==0.4.0
google-auth==2.19.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.3
Gymnasium==0.26.3
gymnasium-notices==0.0.1
h5py==3.8.0
idna==3.4
imageio==2.30.0
importlib-metadata==6.6.0
importlib-resources==5.12.0
jax-jumpy==1.0.0
jsonschema==4.17.3
keras==2.9.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
lazy_loader==0.2
libclang==16.0.0
lz4==4.3.2
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
mdurl==0.1.2
msgpack==1.0.5
networkx==3.1
numpy==1.24.3
oauthlib==3.2.2
opencv-python==4.7.0.72
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.2
PettingZoo==1.22.1
Pillow==9.5.0
pkgutil_resolve_name==1.3.10
platformdirs==3.5.1
protobuf==3.19.6
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pygame==2.1.0
Pygments==2.15.1
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.3.5
rsa==4.9
scikit-image==0.20.0
scipy==1.9.1
six==1.16.0
tabulate==0.9.0
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6
tensorflow==2.9.0
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.32.0
tensorflow-probability==0.17.0
termcolor==2.3.0
tifffile==2023.4.12
typer==0.9.0
typing_extensions==4.6.2
tzdata==2023.3
urllib3==1.26.16
virtualenv==20.21.0
Werkzeug==2.3.4
wrapt==1.15.0
zipp==3.15.0
ray==2.5.0
